# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "HealthController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Estado general del servicio",
        "tags": ["health"]
      }
    },
    "/health/live": {
      "get": {
        "operationId": "HealthController_live",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Liveness del proceso",
        "tags": ["health"]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "HealthController_ready",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Readiness de dependencias",
        "tags": ["health"]
      }
    },
    "/events": {
      "get": {
        "description": "Ejemplo rápido: `GET /events?from=2026-06-11T00:00:00Z&to=2026-06-18T23:59:59Z`",
        "operationId": "EventsController_list",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inicio inclusivo del rango en formato ISO 8601",
            "schema": {
              "example": "2026-06-11T00:00:00Z",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Fin inclusivo del rango en formato ISO 8601",
            "schema": {
              "example": "2026-06-18T23:59:59Z",
              "type": "string"
            }
          },
          {
            "name": "phase",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["SCHEDULED", "LIVE", "FINISHED", "POSTPONED", "CANCELLED", "SUSPENDED"]
            }
          },
          {
            "name": "timezone",
            "required": false,
            "in": "query",
            "schema": {
              "default": "UTC",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lista eventos por rango",
        "tags": ["events"]
      }
    },
    "/events/{eventId}": {
      "get": {
        "operationId": "EventsController_detail",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Detalle y mercados de un evento",
        "tags": ["events"]
      }
    },
    "/betslip/calculate": {
      "post": {
        "operationId": "BetslipController_calculate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculateBetslipDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Calcula un cupón sin generar movimientos",
        "tags": ["betslip"]
      }
    },
    "/betslip/place": {
      "post": {
        "operationId": "BetslipController_place",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculateBetslipDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirma una apuesta con débito atómico",
        "tags": ["betslip"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Inicia sesión y entrega un JWT",
        "tags": ["auth"]
      }
    },
    "/balance": {
      "get": {
        "operationId": "BalanceController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["balance"]
      }
    },
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["notifications"]
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_read",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["notifications"]
      }
    }
  },
  "info": {
    "title": "API Mundial 2026",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
      "CalculateBetslipDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["SINGLE", "COMBO"]
          },
          "stake": {
            "type": "string",
            "example": "10.00"
          },
          "selectionIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["type", "stake", "selectionIds"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "demo@apuestatotal.test"
          },
          "password": {
            "type": "string",
            "example": "Demo12345!"
          }
        },
        "required": ["email", "password"]
      }
    }
  }
}